fix(setup): update rest request utils function
