+++ knip.json
@@ -1,13 +1,11 @@
 {
     "$schema": "https://unpkg.com/knip@5/schema.json",
     "entry": [
-        "src/**/*.test.ts",
         "src/**/*.test.tsx",
         "src/index.tsx!"
     ],
     "project": [
         "src/**/*.d.ts",
-        "src/**/*.test.ts",
         "src/**/*.test.tsx",
         "src/**/*.tsx!",
         "src/**/*.ts!"
