+++ package.json
@@ -9,7 +9,7 @@
         "test": "vitest",
         "coverage": "vitest run --coverage",
         "typecheck": "tsc",
-        "lint": "pnpm lint:js && pnpm lint:css",
+        "lint": "pnpm lint:js && pnpm lint:css && pnpm lint:unused",
         "lint:fix": "pnpm lint:js --fix && pnpm lint:css --fix",
         "lint:js": "eslint src",
         "lint:css": "stylelint \"./src/**/*.css\"",
@@ -63,7 +63,6 @@
         "stylelint-no-unused-selectors": "git+https://github.com/toggle-corp/stylelint-no-unused-selectors#e0831e1",
         "stylelint-value-no-unknown-custom-properties": "^6.0.1",
         "typescript": "^5.5.2",
-        "typescript-eslint": "^8.26.0",
         "vite": "^5.1.5",
         "vite-plugin-checker": "^0.6.0",
         "vite-plugin-compression2": "^1.0.0",
