+++ src/App/index.tsx
@@ -4,9 +4,9 @@ import {
     createBrowserRouter,
     RouterProvider,
 } from 'react-router-dom';
-import { RequestContext } from '@togglecorp/toggle-request';
 
 import RouteContext from '#contexts/route';
+import { RequestContext } from '#utils/restRequest';
 import {
     unifiedError,
     unifiedOptions,
