+++ src/utils/restRequest/index.ts
@@ -1,26 +1,25 @@
 import {
-    LazyRequestOptions,
     RequestContext,
     RequestOptions,
-    useLazyRequest,
     useRequest,
 } from '@togglecorp/toggle-request';
 
+import { NotesList } from '#utils/types';
+
 import { TransformedError } from './unified';
+
+type NotesResponse = {
+    page: number;
+    page_size: number;
+    results: NotesList[];
+}
 
 /** @knipignore */
 // eslint-disable-next-line max-len
-const useUnifiedLazyRequest: <R, C = null>(requestOptions: LazyRequestOptions<R, null, C, null>) => {
-    response: R | undefined;
-    pending: boolean;
-    trigger: (ctx: C) => void;
-    context: C | undefined,
-} = useLazyRequest;
-
-const useUnifiedRequest: <R>(
-    requestOptions: RequestOptions<R, TransformedError, null>
+const useUnifiedRequest: (
+    requestOptions: RequestOptions<NotesResponse, TransformedError, unknown>
 ) => {
-    response: R | undefined;
+    response: NotesResponse | undefined;
     pending: boolean;
     error: TransformedError | undefined;
     retrigger: () => void;
@@ -28,6 +27,5 @@ const useUnifiedRequest: <R>(
 
 export {
     RequestContext,
-    useUnifiedLazyRequest as useLazyRequest,
     useUnifiedRequest as useRequest,
 };
