+++ src/views/Root/index.tsx
@@ -2,7 +2,6 @@ import { Outlet } from 'react-router-dom';
 
 import styles from './styles.module.css';
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     return (
